finished making code compatible with mac
still missing some dependencies
had to rewrite RE_ASSERT
also forgot the tgui headers…heh…

+++ REngine3/source/re/graphics/sw/Starfield.cpp
@@ -133,7 +133,8 @@ void Starfield::setColor(const sf::Color color)
 void Starfield::draw(sf::RenderTarget& target, sf::RenderStates states) const
 {
 	states.transform = getTransform();
-	const unsigned int size{ m_vertices.size() };
+    // changed to static cast by REngine3. This is to silence errors on XCode 8.2.1
+	const unsigned int size{ static_cast<unsigned int>(m_vertices.size()) };
 	if (size > 0)
 		target.draw(&m_vertices.front(), size, m_primitiveType, states);
 }
